Bug#30485709 "TESTNODERESTART -N TWONODEFAILURE T6 T13" FAILS (.4NODE4RPL)
Many node restart tests rely on assumptions that are not valid in the
context of 3- and 4- replica configurations.

This patch adds several new methods to class NdbRestarter, and then
attempts to use them in testNodeRestart.ccp and NdbRestarts.cpp so
as to correct the invalid assumptions for some failing tests.

 - getNumNodeGroups() returns the number of node groups
 - getNumReplicas() returns the number of replicas alive in at least one
   group
 - getMaxConcurrentNodeFailures() returns the number of nodes that can
   simultaneously fail in a test -- generally half the number of data nodes.
 - getMaxFailedNodes() returns the total number of nodes that can fail,
   calculated as (replicas - 1) * node groups. For 3- and 4- replica
   configurations, this number cannot be reached by a single node failure
   event, but only through a series of events.
 - getRandomNodePreferOtherNodeGroup() first attempts to provide a random
   node from another node group, then falls back on a random node from
   the same group. Using this in place of getRandomNodeOtherNodeGroup()
   can allow an existing test to run in new configurations, albeit with
   slightly changed test semantics.

For the tests declared in NdbRestarts.cpp, the test decalaration is
expanded to include the number of node groups required for running the
test. The function in testNodeRestart.cpp that runs these tests is
renamed to runNamedRestartTest(), to help clarify that the execution
path depends on the test name.

Many of the fixed tests are also changed to return NDBT_SKIPPED as
appropriate rather than NDBT_OK.

This patch consolidates two previous patches for this bug and supercedes
a previous version.

Change-Id: Ib263b2c6b3554c7d13791b40587e10fbc367e49c
Reviewed-By: Sanjana <sanjana.ds@oracle.com>