Bug#30477993: MYSQL CRASH AFTER ATTEMPT TO RUN SELECT QUERY
Queries that used the MEMBER OF operator could make the server exit.

A check for whether an index could be used for the expression tried to
access a member of the collation returned by
Item::compare_collation(), which was nullptr. For most Item types,
compare_collation() returns nullptr. The fix is to reorder the checks
so that the member is only checked for Item types where it is known
that compare_collation() != nullptr.

Change-Id: I27772c4015071b489f4ff92733d87a46e65c6d42