Bug#30529220 NDB_CONFIG --XML --CONFIGINFO CHANGED DISPLAY ORDER OF VALUES FOR ARBRATION
Restore the order of enumerated values as shown

  ndb_config --xml --configinfo

to the order used in 8.0.17 and earlier.

In 8.0.18 the order was accidentally changed as part of WL#13143.

That made it impossible to deduce how to interpret the numerical value that
ndb_config and ndb_mgmd --print-full-config shows for the enumerated
cluster configuration parameters like Arbitration and
DefaultOperationRedoProblemAction.

Reviewed-by: Sanjana DS <sanjana.ds@oracle.com>