BUG#30498898: Fix problems with Prepare LCP in PGMAN
Reviewed-by: Mauritz Sundell <mauritz.sundell@oracle.com>

Prepare LCP handling was introduced by WL#12924 in 8.0.19 to ensure a
smooth writing of disk data pages for checkpoints even in the presence
of extremely challenging workloads. Checks were put in place to
ensure that no Prepare LCP writes was still outstanding when a checkpoint
completed.

These checks failed in ndbd due to that Prepare LCP activity was
activated when it was only intended to perform a local synch of
the extent pages, in ndbd the same instance is used for extent
pages and data pages.

It also failed when a table was dropped, in this case a table
could be removed from checkpointing but still perform Prepare LCP
writes. In order to avoid this conflict a new phase at
Prepare Drop table in LQH was added that ensures that we stop
handling Prepare LCP for tables being dropped, we wait until all
outstanding Prepare LCP writes for the table have been processed
before we send the response back of PREP_DROP_TAB_REQ from LQH.

A similar problem also occurs when a table is created and we
start performing PREP_LCP writes and the fragment isn't checkpointed
in the current LCP.

To support this a new table record was created in PGMAN.