WL#7636: Extend SCAN_FRAGCONF and SCAN_TABCONF signals
Introduce extensions to the SCAN_FRAGCONF and SCAN_TABCONF signals where an extra
word containing a bitmap with the SPJ nodes / tables still having more rows to be
returned for this batch of result rows.

For the SCANTAB_CONF handling version control needed the ndbd_send_active_bitmask(version)
method to be added. This enables an axtra 5'th word to be send/received between TC and
the NdbQuery API when both has version > 8.0.20. Else the old 4 word format is used.

For SPJ -> TC FRAGCONF the version is simpler: An updated SPJ will always send the
extra activeMask word added by this patch. If this signal is receive by a
too old TC, the extra word will simply be ignored. If an updated TC receive a
FRAGCONF which are not of the extended format, it will set activeMask = 0.

Note that such activeMask's will (later patches) only be required by the SPJ API
doing result set handling of outer joined scan tables. As we did not
push such queries prior to this WL, just ignoring the non existing activeMask
should be fine.

Reviewed-by: Mikael Ronström <mikael.ronstrom@oracle.com>