WL#7636: Fix minor bug in SPJ query planning:
The nextBranchPtr argument to Dbspj::planSequentialExec() should be a
NullTreeNodePtr when appending treeNodes for scans.

Dbspj::planSequentialExec is called two different places for a scan node.
For one of these places a non-null treeNode argument was used for the
nextBranchPtr.

We believe that this bug will only be a problem when (later) starting to
push outer joined scan (-branches). It was found and fixed as part of
extensive RQG testing of later patches in this patch series.

Reviewed-by: Mikael Ronström <mikael.ronstrom@oracle.com>