BUG#30517172: BACKPORT BUG#30517160 TO 5.7
When a member joins to a group, it will go through a recovering
stage on which it updates its state with the group one.
If a error happens on this stage, the new member will retry X times.
The error that happened on each recovery retry is displayed on
performance_schema.replication_applier_status_by_worker table,
though this error was only being displayed when replication applier
was configured to have parallel workers. If the replication applier
was configured as a single thread there was no error reporting.
The same thing was happening on SHOW SLAVE STATUS, though here is
was independent of the replication applier configuration.
This was caused by a internal RESET SLAVE being executed after each
retry attempt, which was cleaning the error reporting.

To solve the above issues, the internal RESET SLAVE after each retry
attempt was removed.
There is no need for it, because it is already done before each
retry attempt.

This is the backport of the fix provided by BUG#30517160: ERRORS NOT
SHOWING ON PS.REPLICATION_APPLIER_STATUS_BY_WORKER ON SINGLE
APPLIER, which by dependency includes the fix provided by
BUG#27940732: CONNECTION ERROR MESSAGES NOT SHOWN IN GR.
GR_RECOVERY_CONNECTION_OPTIONS WRONG.

ReviewBoard: 23340