Bug #29770705: SERVER CRASHED WHILE EXECUTING SELECT WITH SPECIFIC                WHERE CLAUSE
Description:
============
If a condition ANDed with another condition having multiple ORs
results in an impossible range, then server exits trying to use the
ANDed condition next time as it was deleted already.

ANALYSIS:
=========
For the query in bugpage, when t.b=4000 gets ANDed with t.b = 10, it
leads to an impossible range which results in deletion of the node in
the tree. However the same condition needs to be ANDed with another
condition as part of the OR. This leads to assertion as the tree is
already deleted. The assumption for deletion is that, if the node was
needed, it would have been cloned already. However, cloning happens
(in key_and) only when the use_count of the node is higher than "1"
which is not true for this case.

SOLUTION:
=========
In 5.7 and below, use_count handling is not always correct. As this
has completely changed in 8.0, we are not touching use_count handling
to fix this problem. Instead we use the "clone_flag" to determine if
a key needs to be cloned. "clone_flag" is set to true if the
conditions which need to be ANDed (in tree_and) are not simple (has
many OR conditions within). This information is passed to key_and
and other helper functions. Use the same along with use_count check,
to clone a key tree.

Change-Id: Iaf7c8d9ff11e2f328ef31376bc19211af7d02a9b