Bug #30521577: REUSEEXISTING FLAG IS NOT VALIDATED
Description:
It is expected that the reuseExisting flag be validated in the plugin.
It is not. An error is being returned that the option is unknown.
The clients want to pass an optoin called reuseExisting with possible
values true or false. This option would semantically be the same as
create table if not exists.

RB: 23359
Reviewed-by: Lukasz Kotula <lukasz.kotula@oracle.com>