WL#7636:  Fix incorrect rejection of join pushdown caused by the 'current_parents' usage.
is_pushable_as_child() collected the set of parent tables referred by the
fields referred by the REF-keys into the map 'current_parents'. In addition
it used ::is_field_item_pushable() to find the full set of parents tables
which may be used through the equality set.

An example of the quality set usage is the query:

select *
  from t1
  join t1 as t2 on t2.pk = t1.pk
  join t1 as t3 on t3.pk = t2.pk;

Table t3 is joined by 't3.pk = t2.pk'. However, we also know from the t2 join
that 't2.pk = t1.pk'. Thus, we will have an equality set containing [t1.pk,t2.pk,t3.pk].
During optimization the optimizer may use the equality set to rewrite queries,
such that the above query may be represented to the SPJ plugin as if it had
been written:

select *
  from t1
  join t1 as t2 on t2.pk = t1.pk
  join t1 as t3 on t3.pk = t1.pk;   << t2.pk replaced with t1.pk

Analyzing this field_parent relationship for this query, is_pushable_as_child()
will find that for t3:
  t3.current_parents = [t1]    < As optimizer replaced t2.pk -> t1.pk
  t3.field_parents = [t1,t2]   < By equality set.

As 'current_parents' was checked when determining join pushability for t3, we would
incorrectly conclude that t3 was not pushable if t1 was not an ancestor
in the pushed join - Even if the pushed t2 was an alternative field source
available in the equality set.

This patch entirely remove the calculation and usage of 'current_parents', and
replace it with similar logic using 'extend_parents', which contain the 'sum'
of all available field_parents.

A few queries in the ndb_join_pushdown_*.test which were incorrectly rejected
then becomes pushable.

Also slightly change the EXPLAIN_NO_PUSH-text being printed in case
where push is being rejected due to this - To the better I think.

Reviewed-by: Mikael Ronström <mikael.ronstrom@oracle.com>