Bug#30414514 MYSQL-8.0: TABLES ARE NOT POPULATING AFTER NDB RESTORE FINISHED
Problem:
--------

This is a follow-up fix to that of Bug#30000202. The original fix
caused the auto sync mechanism to fail the synchronization of tables
should the SDI definition not match the NDB Dictionary definition of
the table. This solution was designed primarily for cases where the
ndb_restore tool is used in conjunction with the --disable-indexes
and --rebuild-indexes options.

However, this strict behaviour had an adverse effect on the general
scenario when ndb_restore is used to restore metadata. The metadata
restore first creates the tables and then their corresponding indexes.
Should the table be picked up for synchronization in this gap between
table and index creation, the sync of the tables could be unsuccessful

Fix:
----

Relax the comparison logic between the SDI and NDB Dictionary
definitions for the specific case of mismatch in terms of indexes.
Such cases are now treated as temporary errors with the table being
detected again in the next round of change detection

Change-Id: I8952b508f59e1b248e48014fdbe9a81894d36e18