Bug #29415012 BACKUPLOGBUFFERSIZE DOES NOT MATCH USER DEFINED VALUE
Revert a code change from WL#9638 which removes the read of the
BackupLogBufferSize config value. If the BackupLogBufferSize is
configured, the backup log buffer size is set to the configured value.
If this parameter is not set, the backup log buffer size is set to
the default value of this config parameter, which is 16MB.

For config parameters which are deprecated, the "description" field
should contain the name of the alternative configuration parameter
to be used. Updated the descriptions of the deprecated parameters
to specify that hardcoded values will be used instead of these
parameters.

Modified the deprecation error message to make it clear that the
deprecated parameter is being ignored.

Approved by : Maitrayi Sabaratnam <maitrayi.sabaratnam@oracle.com>

Change-Id: I1cb066b45ca8d9756af194c219051d764483e004