WL#7636, Rewrite ha_ndbcluster_push.cc to use the join_nest structure.
The join-nest (and sj_nests) are what the optimizer itself use in order to
provide information about the inner/outer join relationship between tables.

When allowing outer joining of scan tables (in later patches), the
current simplified approach becomes insufficient to fully determine
how tables are nested inside joins.

Thus this patch introduce the join_nests structure to the join pushability
anylyzer in ha_ndbcluster_push.cc.

The existimg method Table_access::get_join_type() is deprecated.

The patch itself is rich on comments on how the join_nests are
represented and used, so will not repeat that here.

Reviewed-by: Mikael Ronström <mikael.ronstrom@oracle.com>