BUG#30501414: SUMA HANDOVER CAN COMPLETE SEVERAL TIMES WITH 3 REPLICAS
Reviewed-by: Frazer Clement <frazer.clement@oracle.com>

Crash was in ndbrequire() around DIH restart progress tracking
code, which observed 'SUMA handover complete' twice, and did not
expect it.

In fact the DIH code was triggered by the unlock of a DICT lock called
'SumaHandover', which was being unlocked twice.

This happened because with 3 (or 4) replicas, a starting node may need
subscription handover from 2 (or 3) different nodes.  These handovers
occur on gci boundaries, and may occasionally occur on 2 (or 3) different
gci boundaries.

The SUMA code for unlocking the SumaHandover DICT lock was sending
one DICT_UNLOCK_ORD for each completed handover, so could send
2 (or 3) unlocks to DICT resulting in this crash.

Fixes
  - Root cause : Modify SUMA to only send DICT_UNLOCK_ORD when
    all handovers have completed
  - All nodes involved in handover have responded
  - All handover gcis have been passed and handovers have
    completed
  - DICT unlock : Modify DICT to only allow a lock to be unlocked
    once
  - DICT unlock : Modify DICT to log the lock name correctly to
    aid debugging

A new testcase testNodeRestart -nSumaHandover3rpl is added.