Bug#30360695: TABLE WITH SUBPARTITIONS CAN NOT ADD COLUMN AFTER UPGRADE
Problem: When upgrading a database with a subpartioned table from
8.0.16 or earlier and then executing ALTER TABLE ADD COLUMN, an assert or
error would occur. Root cause was that a bug fixed in 8.0.17 prevented
sdis from being stored for subpartitioned tables. An instant ALTER
after upgrade would then fail because the missing sdi could not be
deleted.

Solution: Trap missing record errors when attempting to delete sdis,
log a warning about it, and report success.

Change-Id: I7a3358b01500342131807e55c9e1b9df4dff1b82