Bug #30216144 MYSQL 8.0 WL#12757 : AUTOTEST TESTUPGRADE NOT AWARE OF NEW LIMITATIONS
WL#12757 removed support at the transporter level for connections between nodes
at versions < 7.4, and nodes at versions > 8.0.

Naturally this causes lots of upgrade + downgrade tests covering those versions
to fail.

Arguably the WL should have added a test to *check* that connections were cleanly
refused (e.g. passes when the feature is pushed rather than fails).

This is fixed by adding a check whether the limitations of WL#12757 mean that
a test should be skipped.
This is done in a new version check as this code is needed from 7.2 up  - and
the existing check named 'ndbd_upgrade_ok()' exists only in 8.0+.

Comments are added to the ShowVersions testcase in an attempt to document that
it should not have any extra 'cluster work' added, and purely work to capture
version info.

Approved by : Sanjana DS <sanjana.ds@oracle.com>