Bug#30525561 FORMAT_PICO_TIME() RETURNS NULL FOR SUBSEQUENT ROWS AFTER BEING CALLED WITH NULL
Problem:

When the native function format_pico_time() is invoked to
process several rows, once a NULL argument is found in a row,
every result after that is set to NULL.

Root cause:

The code contains the following sequence:
- a check for args[0]->null_value
- a call to args[0]->val_int()
in that order.

This is incorrect, the argument needs to be evaluated first,
with val_int(), before looking at the null_value.

The format_pico_time function was in fact using the null_value
of the previous record and not the current record,
falling in a "forever return null" mode once null is found.

Fix:

Use the proper order, that is:
- evaluate function arguments first
- check for argument nullity / type / sign / values second.

The native functions:
- format_bytes()
- ps_thread_id()
had the same bug, following the same code pattern,
and are fixed as well.

Approved by: Chris Powers <chris.powers@oracle.com>