Bug#30556487 MYSQLD HANG IN NDB_STORED_GRANTS CODE ON CREATE USER
In ndb_stored_grants, when write_snapshot() attempts to update a
user record, it begins by reading the record. If the record is
not found, the read operation returns a 626 error, and this
error was not handled correctly. Correct the error handling.

Also correct the error handling from ndb_trans_retry, so that
temporary errors are written to the MySQL server log file at
info level, but errors that resulted in the failure of the
transaction are written at error level. (This originated
as a patch for bug#30536124).

Reviewed-by: Lakshmi Narayanan Sreethar <lakshmi.narayanan.sreethar@oracle.com>