Bug#30576983 ASAN DETECTED MEMORY LEAK IN NDB_DDL.ALTER_OPEN_TRANS
The fix for Bug#29911440 aborts an ongoing transaction if the table
definition has changed midway. When the transaction is aborted, it
fails to free the malloc'd data memory returned by the getExtraMetadata
method. Due to this, when the ndb_ddl.alter_open_trans test is run in
an ASAN build, it reports a memory leak. Fixed that by properly freeing
the data memory before aborting the transaction.

Change-Id: Ied99e949f712cf219fbcf2650b77116d6dc2840d