Bug#29630767 - USE OF UNINITIALIZED VALUE IN LIBMYSQL                (CLIENT.CC FUNCTION RUN_PLUGIN_AUTH)
DESCRIPTION
===========
The authentication plugin name is not properly initialized
if some of the fields information in the connection packet
is incorrectly populated.

FIX:
====
In such case, we set the values of the fields in such a way
that the default plugin gets used for authentication.
The corresponding data associated with the original plugin
is also discarded.

RB: 23301