Bug #30354273 SUPPORT IPV6 ADDRESSES FOUND IN THE METADATA.
Refactored test_state_file to run faster (28s -> 5s)
and be more reliable. Instead of the arbitrary sleeps
before doing checks, used check-sleep(step) loop strategy
for verifying state-file and logfile contents.

Fixed handling of the IPV6 if it appears in the metadata.
Fixed the mock server to accept the connections to IPV6 address.

RB: 23397