BUG#30445128 `PRIVILEGE_CHECKS_USER` MUST NOT FORCE `REQUIRE_ROW_FORMAT` TO BE SET
Description/Analisys
--------------------

With WL#12966, `REQUIRE_ROW_FORMAT` was introduced as an option for `CHANGE
MASTER TO`. The implementation bound the value of `PRIVILEGE_CHECKS_USER` to the
value of `REQUIRE_ROW_FORMAT`, making it mandatory for it to be `1` for any
value of `PRIVILEGE_CHECKS_USER` different from `NULL`. This would remove the
hability to use `PRIVILEGE_CHECKS_USER` with statement-based-replication and
would remove a feature within a GA version.

So, the link between the two options must be removed.

Fix
---
1) Remove the checks for mandatory `REQUIRE_ROW_FORMAT` while setting
   `PRIVILEGE_CHECKS_USER`.
2) Remove the checks for instantiated `PRIVILEGE_CHECKS_USER` when unsetting
  `REQUIRE_ROW_FORMAT`.
3) Change test infra-structure and tests that check and up-hold the said
   dependency.

RB: 23376
Reviewed-by: Pedro Gomes <pedro.gomes@oracle.com>
Reviewed-by: Luís Soares <luis.soares@oracle.com>