Bug#30578557: REMOVE OBSOLETE WORKAROUNDS FOR GCC BUGS AND WARNINGS
Remove workarounds that were required for versions of GCC we no
longer use. This includes workarounds for GCC bugs and GCC warnings.

Of particular note is the use of 'volatile double' to force
GCC to flush double values out of 80-bit FPU registers before
doing computations. This workaround is not required on 64 bit
since SSE registers are used there. The workaround is not
required on 32 bit either because mysqld (since 2009) at startup
explicitly sets 32 bit FPU precision to double precision for
portable results of floating point operations.

Change-Id: Ia8eeef7309b95e1f83aaec960759e0668607c834