Bug#30503085 - NDB_CONDITION_PUSHDOWN FAILS WITH RESULT CONTENT MISMATCH
Background: The statistics 'Ndb_api_read_row_count' shows the number
of rows a query retrieved from Ndb. This is counted when a TRANSID_AI
signal is received by the API.

Problem: ndb_condition_pushdown.test fails with 'result content
mismatch' because it expects that the statistics 'Ndb_api_read_row_count' from
performance_schema.global_status after the query 'select count(*) from
t1;' should show the table's total #rows.

Though this expectation is true for queries retrieving the qualified
rows via a scan, the query 'select count(*) from t1;' is an
exception. To make it efficient, only pseudo rows are retrieved, where
one row contains a fragment_id and the #rows it contained, and having
the following format:
< AttributeHeader::FRAGMENT, AttributeHeader::ROW_COUNT>.

Solution: The purpose of the test is to create a baseline with the
number of rows of a table. This can be achieved by selecting all the
rows.

Reviewed-by: Ole John Aske <ole.john.aske@oracle.com>