Bug #30371319 CLUSTER_ID MISMATCH ON SLAVE AND PRIMARY DURING BOOTSTRAP IGNORED
When bootstrapping against non-master node, when we fallback to
the other nodes we now check if the current node has the same
cluster_id (for ReplicaSet cluster) or replication_group_name
(for GR cluster). If it does not we skip it and do not try to
bootstrap from it.

RB: 23426