Bug #30478016 ROUTER NOT WORKING PROPERLY WITH --CONF-USE-GR-NOTIFICATIONS
The Router did not handle properly the scenario when x plugin
version does not supported setting mysqlx_connect_timeout variable.
In that case it would not set up the GR notifications
properly. This patch changes this behaviour and now not being
able to set it is not an error, there is only a info log about it
and we still try to setup the GR notifications listeners.

Another matter is that once there is no mysqlx_connect_timeout
support there likely also is no GR notifications support.
For that scenario the error reporting got improved.

Added component tests for both those scenarios.

RB: 23313