Bug#29995256: PRIVILEGE_CHECKS_USER NOT CLONED FOR RECOVERY CHANNEL Bug#30510766: SLAVE FAILS TO START AFTER BEING CLONED IF USING POSITIONS
Currently when a slave or group replication member is cloned, the
settings for a channel privileged user or the require row format are
not cloned.

To fix this we now enable the cloning of the the slave_relay_log_info
and slave_worker_info tables now making the clone completly agnostic
to replication data.

One issue however is that on start there is information about relay
log files that no longer exists and even without this, until now the
new cloned slave would start to replication from scratch if auto
positioning was not enabled.

To fix this we now try to execute relay log recovery over the cloned
slave info data. We also output a warning telling the user that in
case this recovery process fails some manual steps to adjust
replication are needed.

ReviewBoard: 23401