Bug#30579922 INCORRECT RESULT IF HASH-JOIN ALGORITHM EXHAUST ITS JOIN_BUFFER
If the hash join build input was a nested loop with a filter
on the inner side and the hash join was not allowed to spill
to disk, we could end up returning the wrong result to the
client. Consider the following execution tree:

        Hash Join
        |       |
  Nested loop  t1
  |         |
  t3    Filter: (t3.i < t2.i)
               |
              t2

If a re-fill of the hash table was triggered while the nested
loop iterator was in the state "reading from inner side", the
filter condition would look at the row in t3's record buffer.
The row in t3's record buffer would not necessarily be the last row
returned from the storage engine, since the probe phase may load
any row in the hash table into the record buffers. This caused
the filter to look at the wrong data, which again could lead to
wrong results. To fix this, the hash join iterator will now
restore the last row read from the build iterator before doing
re-fill of the hash table.

Change-Id: Icdd43025fbacc69c5eb6986924f27d2022970095