Bug #30417719 ASSERTION FAILURE: FSP0FSP.CC:2371:N_USED_NOT_FULL > 0 || N_TOTAL_NOT_FULL == 0
Problem:

The original fix for this bug at rb#23349 has an issue.  When the persistent
cursor is stored and restored, the position of the record could change. Also,
when the purge thread does an intermediate commit, it is possible that the
delete marked record is re-used.

Solution:

After persistent cursor is restored, reset the LOB reference to the correct
value.  Also, the callers need to be aware that the cursor position could be
changed.  Check if the clust_rec is re-used by checking the delete mark flag.

rb#23459 approved by Kuba <jakub.lopuszanski@oracle.com>