Bug #30444982 : BUG #30427675 WL#8426 : "TESTDICT -N BUG13416603 I2" FAILS
Problem
-------
Function do_action() assumes that the nodelist containing the index stat nodes
is ordered but the nodelist is not always ordered in the same way in all the nodes.
The order could change if one or more data nodes restart.
Hence, do_action() wrongly ignores an index stat update request in this case.
Therefore, all the nodes ignore the request resulting in no update made to the
index statistics tables.

Solution
--------
-> Make sure that the node list is ordered in the same way in all the nodes.

Approved-by: Mauritz Sundell <mauritz.sundell@oracle.com>