Bug#29859605 INVALID READ OF SIZE 1 IN SOME PFS TESTS
When the performance_schema.users table is truncated, the statistics
for each user are aggregated across all instrumented threads.

The stack in the bug report shows the invalid access of a THD object
during the aggregation of an instrumented thread, but the corresponding
THD object has apparently been freed, indicating that the thread
instrumentation was orphaned.

In the internal THD api, create_thd() calls pfs_new_thread() when
a THD is created, but does not call pfs_delete_thread() when the
THD is deleted.

This patch adds pfs_delete_thread() to delete_thd() to ensure
that the thread instrumentation is removed when a THD is deleted.

Approved by: Marc Alff <marc.alff@oracle.com>