Bug#29539976 LOCK_ORDER: CONNECTION_CONTROL PLUGIN   INSTRUMENTATION IS NOT DISCOVERABLE
Before this fix:

After loading the connection_control plugin,
the performance schema instrumentation embedded in the code
was not visible in table performance_schema.setup_instrument.

This is a problem because available instruments are harder to discover:
the workload has to execute a specific code path for the instrument
to be eventually added in the performance schema.

In this case, a connection had to actually be blocked
by the connection control plugin for the mutex and cond to appear.

This problem in particular affects the LOCK_ORDER tool,
which failed to find the instrumentation in the
connection control plugin.

Root cause:

mysql_mutex_register() was not called on plugin init
but later during the plugin execution.

Fix:

The performance schema instrumentation
for the connection_control plugin
was reviewed in details, leading to the discovery
and fix for the following issues:

1)

Mutexes, rwlocks and conditions were not registered up front
on plugin init, but later.

For mutexes and conds, the registration was executed
multiple times, which is un necessary.

For rwlocks, the registration was correct (done once and always),
but the code pattern was inconsistent with instrumentation in general.

All items are now registered once at plugin init.

2)

A stage was defined, but never registered,
so that the performance schema was just ignoring it,
treating the stage as un instrumented.

The stage is now instrumented properly.

3)

Mutex connection_delay_mutex was instrumented as a singleton,
even when the real mutex is not.

Fixed by removing the PSI_FLAG_SINGLETON flag.

4)

Rwlock connection_event_delay_lock was a singleton in the code,
but not instrumented as such.

Fixed by adding the PSI_FLAG_SINGLETON flag.

Approved by: Chris Powers <chris.powers@oracle.com>