Bug #30561982: SET PERSIST DOES NOT WORK ON WINDOWS OUT OF THE BOX RB#23464
mysqld has a bunch of global string variables to hold various
"important" paths, like e.g. the mysql "home" directory, the mysql
"data"  directory etc.
Calculating these is a multi-step process.

For the data directory, if you do not specify one, it's "calculated" as
follows:
1. the program name (argv[0] is expanded to a full name).
2. the data directory was initially calculated as the relative directory
between the hard-coded "home" directory and the hard coded "data"
directory. This usually is the relative directory name "data".
3. the "home" directory (mysql_home) is calculated as the directory the
program is in.
4. the "data" directory is calculated by adding to the "home" directory
to the already calculated relative directory
5. if any of these are specified on the command line or in the config
file(s) they are read from there, override the calculated ones and are
made absolute paths.
6. command line options are processed

Step 1 and 2 are done real early (at the start of the main() function).
But the subsequent steps were performed relatively late: in fix_paths(),
    before step 6.
Thus, the persisted command line variables extension processing
(part of step 5) doesn't yet have the fully calculated data
directory or mysql_home.
To make things even worse, the persisted handler class was actually
copying the value of the "data" directory even earlier than this:
after step 2. Thus, if no datdir specified on the command line, it
was just trying to expand the "data". And this results in "data"
(for the purposes of finding the persisted config file) being
expanded as a subdirectory of the current directory. And that's
the wrong directory. Thus when you try to save something in the
persisted file it cannot find the actual JSON file and returns an
error.

Fixed by:
isolating step 3 into a separate utility function and calling this
after step 1 (as it solely depends on step 1 and nothing else).
doing step 4 for the cached value of persisted variables cache
when it's initialized after step 2 in case there's no data dir
specified on the command line.

In this way, the "data" directory value cached after step 2 is the
same as the one calculated originally in step 4.