Bug#30132988: FIX WARNINGS FOR PGO + LTO [noclose]
Fix One Definition Rule warning (-Wodr) in XCom when built with LTO.
The problem was that definition of the connection_descriptor struct
depended on if the XCOM_HAVE_OPENSSL symbol was set. It was set for
the GR plugin and some, but not all, unit tests. Since we did not
set it consistently, the linker found the same struct with two
different definitions.

This patch fixes the problem by swiching the #ifdef check from
XCOM_HAVE_OPENSSL to !XCOM_WITHOUT_OPENSSL. Since building with
SSL support is the common case (and always true for the server),
it makes more sense to define a symbol when SSL support is not
present (MGM).

Change-Id: I71ad254ce9a58f2b59148684f20ae7cbb9a3e048